+++ packages/backend/src/apps/postman/__tests__/actions/send-transactional-email.test.ts
@@ -79,6 +79,46 @@ describe('send transactional email', () => {
     )
   })
 
+  it('should throw step error for invalid template (blacklist)', async () => {
+    // simulate postman error
+    const error400 = {
+      response: {
+        data: {
+          code: 'invalid_template',
+          message: 'Recipient email is blacklisted',
+        },
+        status: 400,
+        statusText: 'Bad Request',
+      },
+    } as AxiosError
+    const httpError = new HttpError(error400)
+    mocks.sendTransactionalEmails.mockRejectedValueOnce(httpError)
+    // throw partial step error message
+    await expect(sendTransactionalEmail.run($)).rejects.toThrowError(
+      'Status code: 400',
+    )
+  })
+
+  it('should throw step error for rate limit', async () => {
+    // simulate postman error
+    const error429 = {
+      response: {
+        data: {
+          code: 'rate_limit',
+          message: 'Too many requests. Please try again later.',
+        },
+        status: 429,
+        statusText: 'Bad Request',
+      },
+    } as AxiosError
+    const httpError = new HttpError(error429)
+    mocks.sendTransactionalEmails.mockRejectedValueOnce(httpError)
+    // throw partial step error message
+    await expect(sendTransactionalEmail.run($)).rejects.toThrowError(
+      'Status code: 429',
+    )
+  })
+
   it('should throw step error for currently unavailable service', async () => {
     // simulate postman error
     const error503 = {
